requirements.txt
streamlit
google-generativeai
python-dotenv
mysql-connector-python==8.3.0

